chore: Update TextureSelectorPanel to use Size Panel
