+++ Unity Project/Dungeoneering/Assets/_Libs/DungeonCrawler.Unity.Editor/src/SelectionInfoPanel.cs
@@ -20,7 +20,7 @@ namespace CaptainCoder.Dungeoneering.Unity.Editor
         [SerializeField]
         private UndoRedoStackData _undoRedoStack;
         [field: SerializeField]
-        public DungeonTextureSelectorController TextureSelector { get; private set; }
+        public TextureSelectorPanel TextureSelector { get; private set; }
         [SerializeField]
         private GameObject _content;
         [SerializeField]
