+++ Unity Project/Dungeoneering/Assets/_Project/Prefabs/UI/Texture Selector/TextureSelectorPanel (Canvas).prefab.meta
